+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/cap-config-writer
 
+## 0.9.11
+
+### Patch Changes
+
+-   Updated dependencies [fffc3a7]
+    -   @sap-ux/fiori-generator-shared@0.9.0
+
 ## 0.9.10
 
 ### Patch Changes