+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.0.27
+
+### Patch Changes
+
+-   d6118c9: Changes to support adding CAP MTA prompt to allow user generate MTA
+
+## 0.0.26
+
+### Patch Changes
+
+-   Updated dependencies [fffc3a7]
+    -   @sap-ux/fiori-generator-shared@0.9.0
+
 ## 0.0.25
 
 ### Patch Changes