+++ packages/ui5-test-writer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui5-test-writer
 
+## 0.5.0
+
+### Minor Changes
+
+-   fffc3a7: Add Freestyle OPA templates to ui5-test-writer
+
 ## 0.4.1
 
 ### Patch Changes